pyproject.toml
[project]
name = "SeeQt"
version = "0.1.0"
description = "Performant data viewers implemented in PySide6"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "napari>=0.6.6",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "polars>=1.34.0",
    "pyqtgraph>=0.13.7",
    "pyside6>=6.9.2",
    "pyyaml>=6.0.3",
    "streamlit>=1.50.0",
    "tifffile>=2025.10.16",
]

[project.scripts]
seeqt-video-viewer = "video_viewer.app:main"
